+++ sdk-js/src/lib/span-hooks/hook-http.js
